Enhance CI configuration and testing setup
- Updated CI workflow to run the full lane on every pull request, improving test coverage and stability.
- Modified `test:ci:full` command in `package.json` to streamline testing processes.
- Added Vitest configuration for forked pools to stabilize coverage runs.
- Enhanced documentation in `README.md` and `CHANGELOG.md` to reflect changes in CI processes and testing requirements.
- Introduced new unit tests for Matrix client mocks to improve test reliability and maintainability.
