+++ .github/workflows/ci.yml
@@ -30,7 +30,6 @@ jobs:
     runs-on: ubuntu-latest
     if: >
       (github.event_name == 'push' && github.ref != 'refs/heads/master') ||
-      (github.event_name == 'pull_request' && github.base_ref != 'master') ||
       (github.event_name == 'workflow_dispatch' && inputs.run_full != true)
 
     steps:
@@ -56,8 +55,8 @@ jobs:
     name: Full lane (coverage + full e2e)
     runs-on: ubuntu-latest
     if: >
+      github.event_name == 'pull_request' ||
       (github.event_name == 'push' && github.ref == 'refs/heads/master') ||
-      (github.event_name == 'pull_request' && github.base_ref == 'master') ||
       github.event_name == 'schedule' ||
       (github.event_name == 'workflow_dispatch' &&
       inputs.run_full == true)
