+++ tests/unit/composables/useMatrixClient.rooms.spec.ts
@@ -1,37 +1,16 @@
 import { beforeEach, describe, expect, it, vi } from 'vitest'
-import { buildMatrixSdkMock } from './matrixClientSdkMock'
+import './matrixClientSpecMocks'
+import { createClient } from 'matrix-js-sdk'
+import { initAsync as initCryptoWasm } from '@matrix-org/matrix-sdk-crypto-wasm'
 import {
+  prepareMatrixClientSpecFile,
   setupFreshMatrixClientGlobals,
   setupMatrixClientTestGlobals,
 } from './matrixClientTestSetup'
 
-const matrixMocks = vi.hoisted(() => ({
-  createClient: vi.fn(),
-  initCryptoWasm: vi.fn(async () => undefined),
-}))
-
-vi.mock('matrix-js-sdk', () => buildMatrixSdkMock(matrixMocks.createClient))
-vi.mock('@matrix-org/matrix-sdk-crypto-wasm', () => ({
-  initAsync: matrixMocks.initCryptoWasm,
-}))
-vi.mock('~/utils/videoMetadata', () => ({
-  readVideoMetadata: vi.fn(async () => ({
-    durationMs: 5000,
-    w: 640,
-    h: 360,
-  })),
-  captureVideoThumbnail: vi.fn(async () => (
-    new Blob(['thumb'], { type: 'image/jpeg' })
-  )),
-}))
-
-const createClient = matrixMocks.createClient
-const initCryptoWasm = matrixMocks.initCryptoWasm
-
 describe('useMatrixClient rooms', () => {
   beforeEach(() => {
-    vi.resetModules()
-    vi.clearAllMocks()
+    prepareMatrixClientSpecFile()
     setupMatrixClientTestGlobals()
     localStorage.clear()
     if (typeof sessionStorage !== 'undefined') {
